requirements.txt
pytest
beautifulsoup4~=4.10.0
mysql-connector-python

nltk~=3.7